requirements.txt
opencv-python
pytesseract
Pillow
numpy
requests
sentence-transformers
scikit-learn
pandas
joblib
streamlit     
matplotlib    
